requirements.txt
beautifulsoup4==4.12.3
discord.py==2.4.0
pynacl==1.5.0
python-dotenv==1.0.1
tidalapi==0.8.2